refactor: Edit model package
